* renamed extension to ubl_vufind_auth * version status is now stable * bumped version to 1.0.0 * added license * updated documentation

+++ ext_localconf.php
@@ -1,17 +1,16 @@
 <?php
+if (!defined('TYPO3_MODE')) {
+    die('Access denied.');
+}
+
 const COMPOSER_AUTOLOAD_PATH = PATH_site . 'Packages/Libraries/autoload.php';
 
-if (!file_exists(COMPOSER_AUTOLOAD_PATH)) {
+if (!file_exists(COMPOSER_AUTOLOAD_PATH) && 'BE' === TYPO3_MODE) {
     $msg = t3lib_div::makeInstance('t3lib_FlashMessage', 'Make sure to install zendframework-stdlib via composer under "Packages/Libraries"', 'Composer autoload not found', t3lib_FlashMessage::WARNING);
     return t3lib_FlashMessageQueue::addMessage($msg);
 } else {
     // require composers autoloader
     require_once COMPOSER_AUTOLOAD_PATH;
-}
-
-
-if (!defined('TYPO3_MODE')) {
-    die('Access denied.');
 }
 
 if ('FE' === TYPO3_MODE) {
@@ -20,7 +19,7 @@ if ('FE' === TYPO3_MODE) {
 }
 
 // add the Vufind Authentication Service
-\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addService($_EXTKEY, 'auth', 'UBL\VufindAuth',
+\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addService($_EXTKEY, 'auth', 'LeipzigUniversityLibrary\VufindAuth',
     array(
         'title' => 'Vufind Authentication',
         'description' => 'Authenticates users based on authenticated vufind session',
@@ -31,6 +30,6 @@ if ('FE' === TYPO3_MODE) {
         'os' => '',
         'exec' => '',
         'classFile' => \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extPath($_EXTKEY) . 'Classes/Typo3/Service/Authentication.php',
-        'className' => 'UBL\VufindAuth\Typo3\Service\Authentication',
+        'className' => 'LeipzigUniversityLibrary\VufindAuth\Typo3\Service\Authentication',
     )
 );
